// frontend/src/lib/api/generated/models/TagStatsItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { TagStatsCategoryCode } from '../models/TagStatsCategoryCode';
import { HttpFile } from '../http/http';

export class TagStatsItem {
    'category': TagStatsCategoryCode;
    'name': string;
    'usageCount': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "category",
            "baseName": "category",
            "type": "TagStatsCategoryCode",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "usageCount",
            "baseName": "usage_count",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return TagStatsItem.attributeTypeMap;
    }

    public constructor() {
    }
}